[gtfs] re-org crate for multiple bins
