+++ .github/workflows/checks.yml
@@ -58,10 +58,10 @@ jobs:
       - run: cargo clippy --all-features
       - run: cargo test --all-features
       - run: cargo build --release
-  gtfs_bbox:
+  gtfout:
     defaults:
       run:
-        working-directory: services/gtfs/gtfs_bbox
+        working-directory: services/gtfs/gtfout
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
@@ -78,6 +78,6 @@ jobs:
             services/transitmux/target/
           key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
       - run: cargo fmt --all --check
-      - run: cargo clippy --all-features
-      - run: cargo test --all-features
+      - run: cargo clippy --all-targets --all-features
+      - run: cargo test --all-targets --all-features
       - run: cargo build --release
