+++ services/gtfs/gtfout/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
-name = "gtfs_bbox"
+name = "gtfout"
 version = "0.1.0"
 edition = "2021"
+
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
