+++ services/gtfs/gtfout/data/mock_gtfs_1/calendar_dates.txt
