+++ services/gtfs/gtfout/data/mock_gtfs_1/fare_attributes.txt
