+++ services/gtfs/gtfout/data/mock_gtfs_1/fare_rules.txt
