+++ services/gtfs/gtfout/data/mock_gtfs_1/feed_info.txt
