+++ services/gtfs/gtfout/data/mock_gtfs_1/stop_times.txt
