+++ services/gtfs/gtfout/data/mock_gtfs_1/stops.txt
