+++ services/gtfs/gtfout/data/mock_gtfs_2/calendar_dates.txt
