+++ services/gtfs/gtfout/data/mock_gtfs_2/fare_attributes.txt
