+++ services/gtfs/gtfout/data/mock_gtfs_2/fare_rules.txt
