+++ services/gtfs/gtfout/data/mock_gtfs_2/feed_info.txt
