+++ services/gtfs/gtfout/data/mock_gtfs_2/routes.txt
