+++ services/gtfs/gtfout/data/mock_gtfs_2/stop_times.txt
