+++ services/gtfs/gtfout/data/mock_gtfs_2/trips.txt
