+++ services/gtfs/gtfout/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod geom;
+pub use geom::{Point, Rect};
